Experimental: Semi Auto Tiling Mode
Holding Shift when activating an AppIcon will open a new tiled window
to fill the free screen space. If there is none or if the free space is
ambiguous and the focused window is tiled, halve the focused window.
The spot of the new window is indicated by a top panel icon. The
settings are in the 'Hidden settings' tab and off by default.