New Feature: Fixed Layout...
A Fixed Layout allows the user to dnd a window to an arbitrary predefined
layout when a specific modifier is pressed while moving a window. By
default, it's `Alt` but it can be changed in the `Hidden Settings`. The
Fixed Layout is the 'favorite' layout in the `Layouts` preference page.
It's possible to change this on the fly with the shortcut `Change favorite
Layout`. Some things were already done in the previous commit.