Fix adaptive tiling and favorite layout combo
If the setting 'adapt edge tiling to the favorite layout' in the hidden
settings was enabled while there was no favorite layout set (which is the
default extension state), we always switched to the default 2 column / 4
quarter tiles while we *may* actually want an adaptive tiling. See initial
discussion in the 'Port to GNOME 42' issue. IMO, it also doesn't make
sense to have that setting enabled by default since there is no favorite
layout by default (and there shouldn't be).