Active Hint: Fix a bug for the always hint and make inner border configurable
Since we call show() on the hint with Meta.later_add it's possible for the
hint to be shown from an old focus change while a newer focus update hid the
hint. Reproducer: Have GNOME Terminal be the only window on the workspace
and run `journalctl -fo cat`. Try to close the Terminal. A warning dialog
should popup. Accept it and close the Terminal. The Terminal will close but
the active window hint will still be shown.

Make the inner border size configurable since on Wayland opening a popup in a
GTK4 app will put the window behind the hint for whatever reason... So give
the user the choice between covering rounded corners and this issue.

Another Wayland/Gtk4 popup bug is that the popup will appear behind the hint.