extension: Perform runtime-only settings overriding
The extension is overriding the user settings by writing on them, sadly this
implies various issues because even if we reset them on extension unloading,
a shell crash or mis-behavior could always lead to affect user settings.

To prevent this, we can temporary override Settings backend read vfunc so
that the values we want to override won't ever be written to user db and so
there will be no risk of leaving them in a dirty state.

Sadly, this requires fixes to gjs:
  https://gitlab.gnome.org/GNOME/gjs/-/merge_requests/831