extension: Add SettingsOverrider to preserve user settings
The extension is overriding the user settings by writing on them, sadly this
implies various issues because even if we reset them on extension unloading,
a shell crash or mis-behavior could always lead to affect user settings.

Tiling assistant replaces the user settings when running, but not to loose the
user settings, we need to save them around so that when the extension gets
unloaded they can be restored.

So add a SettingsOverrider class that:
 - Replaces the user settings with the extensions one
 - Remembers what has been changed and saves it locally and in our settings
 - When the extension is unloaded the settings are restored or reset

Now, in case gnome-shell is stopped or the extension has crashed for some
reason, once the extension is reloaded we check if previous settings were
saved and in case they are used to restore the user settings on destruction.