Layouts: Fix position of the search-a-layout popup on multi-monitor setups
While the popup was positioned to be at the center of the current monitor in
absolute coords, it was a child of a widget that spanned only the workArea
(and not the full monitors). So instead make the parent equal the Main.uiGroup
and in turn the global.stage.

Fix https://github.com/Leleat/Tiling-Assistant/issues/241