moveHandler: Fix non-extension maximization restore for GNOME 43
If the grabbed window was maximized in a way not supported by the extension
(for instance by double-clicking the titlebar), we still tried to restore the
window's size because the window may have been tiled before it was maximized.
However, the maximized rect was used to calculate the position after the window
was restored. Instead the unmaximized rect should have been used. Fix this now.
This only affects GNOME 43 since on GNOME 44 we return early. So GNOME 44 is
still affected by a similar issue.