Utils: Fix rounding error in Rect.getUnitAt()
getUnitAt is meant to be used when you need to split a rect into nearly
equally sized 'units' e. g. split the workarea in X-number of columns or
X-number of rows. It was meant to prevent rounding errors resulting in
gaps between rects. All units are equally sized except the last one. The
last unit took whatever space was left... and it sorta worked but it
could have resulted in the last 'unit rect' potentially being just 1 px
wide or high depending on whether the calling rect had an odd or even
size.

For instance, Ubuntu 22.04 has a top bar with a height of 27 px. If you
tried to tile a window to the bottom half on a 1080 px high screen, the
tile would end up 1 px short and you'd leave a gap at the bottom of the
screen.

Instead now check whether the unit rect we are looking for is the last
unit and size it accordingly.

Related https://github.com/Leleat/Tiling-Assistant/issues/261.