Prefs: Clarify window movement modes in prefs
PS. the split into a default mode as a radio group and the other modes
having a dropdown for choosing the modifier is a bit awkward (and it
made the code awkward as well when the default mode also has a modifer
set). It's probably better that each mode gets a dropdown and 'default'
being one option for in the dropdown... Oh well, future stuff.

Closes https://github.com/Leleat/Tiling-Assistant/issues/266