moveHandler: Handle change of window action key
Currently, a conflict of the window action key (Alt or Super) with a
movement mode modifier of T-A is only checked and handled during enable
time. But users may change the window action key while T-A is already
enabled. Handle that case now as well. Previously, the modifier for the
conflicting move mode was switched to the non-window-action-modifier.
But that can be an issue, if that modifier is already taken by a
different mode. So simply disable the modifier for the conflicting mode.

Closes https://github.com/Leleat/Tiling-Assistant/issues/313