moveHandler: Remove restore-window-size-on setting
Ideally, if a tiled window is dragged around, we'd restore the window
size and place the window under cursor when the grab starts. In the past
that didn't work consistently for all windows, so I've added the option
for the window to restore its size after the grab is released as a
workaround. I don't think this is needed anymore, at least from the
manual testing I've done... Let's remove this workaround (and see if
other people still encounter it).