FocusHint: Add a focus hint
This commits  basically integrates https://github.com/Leleat/focus-indicator-prototype
into Tiling Assistant. The other extension was just a prototype in
cooperation with a GNOME Designer. But it is more polished than the
active window hint that is currently in Tiling Assistant. I did some
slight refactoring. I also removed the focus indication when using
swipe gestures since the animation didn't work quite well (and had
a bug during the animation with multiple monitors).

This time however, the focus hint will be disabled by default as I don't
believe it is that useful of a feature anymore, it deviates a bit too
much from the stock GNOME behavior, and it may even be an unexpected
behavior from a users perspective, if they don't know what it is or how
it gets triggered. This is feature is also less tested as they are a lot
of permutations of situations to test (multi-monitor, workspaces on all/
primary display, etc). Additionally, it is put behind the advanced and
experimental settings toggle. Some of the settings for the focus hint
are not exposed in the UI (like the outline border radius, outline)
because I'd like to cut down on the settings. Even putting them behind
the advanced toggle doesn't feel good because eventually the advanced
settings will grow so big that there is at least one setting for someone
out there, which will eventually make the advanced settings 'mandatory'.
I kinda regret exposing so many settings in the prefs UI... oh well.

Fixes https://github.com/Leleat/Tiling-Assistant/issues/306
Fixes https://github.com/Leleat/Tiling-Assistant/issues/222