Settings: Add settings watcher capability
It is fairly common for extensions to tie features to toggles in their
settings. Extensions would listen for a settings change and apply/revert
a feature. This commit adds a convenience method to register a handler
for a settings change and optionally apply the handler `immediate`-ly.
The naming is inspired by Vues watchers.