Injections: Add custom props/methods into Mtk.Rectangle
Previously, there was the wrapper class utility.js::Rect. But now that
there is the Injections singleton, it's easier to just inject all custom
properties and methods into Mtk.Rectangle and use the rect directly.
This is pretty much just a copy/paste of the methods from the wrapper
class. I've only omitted some unused methods.