keybindingHandler: maximize vertically/horizontally without gaps
When maximizing a window vertically or horizontally, we've added gaps to
the window like we normally do to tiled windows. This makes sense if the
window is already tiled. If the window is floating, we shouldn't add
gaps to it and instead should keep the currently visible size.

Fixes https://github.com/Leleat/Tiling-Assistant/issues/349