Timeouts: Add convenience class
Using timeouts in an extension can be somewhat common. For every timeout
you create, you need to make sure that it gets removed when the
extension is disabled to pass EGO's review. Simply returning
`GLib.SOURCE_REMOVE` isn't enough since the extension may get disabled
while the timeout is running. EGO is fairly strict about this. So you
usually need to create a variable to track the timer and remove it on
`disable`. When using timers, it's also common to remove the previously
started timer when a new one starts. This commits adds a convenience
class that automatically takes care of these 2 issues.