fix: improve compatibility with aatws
I was lazy in the past and "extended" GNOME's app switcher so that I
didn't need to write the tiling popup from scratch. Besides the fact
that this was a bad usage, it also caused compatibility issues with
other extensions that modify the app switcher or switcher popup like
aatws.

As a quick fix, I've just pulled in the code from the app switcher and
switcher popup from GNOME Shell into the extension. I only amended some
import statements. It would be a nice idea to rewrite the tiling popup
"properly".

fixes: https://github.com/Leleat/Tiling-Assistant/issues/381