fix: disable animations for tiled windows
This is only a workaround for the linked issue below. Currently, retiling a
tiled window via the Tiling Popup may freeze it (see issue for STR). A better
way to fix this would be to implement custom animations for tiling (instead of
misusing GNOME's existing solution).

related: https://github.com/Leleat/Tiling-Assistant/issues/376