feat: optionally make focus hint outline a border
The outline versions (animated + static) of the focus hint used a widget
that is put behind the window with a solid background color. The
reasoning behind this is to account for different border radii of
windows because if the outline actually uses a border there may be a gap
between the focus hint and the window corners. However, this didn't work
well with transparent windows. This commit adds an option to change the
style of the outline to allow for a border around the window instead of
a solid rect. This is also how the previous focus hint (aka active
window hint) worked. This means there is a gap between the window
corners and the focus hint but that is better than the focus hint being
seen through transparent windows.

Fixes https://github.com/Leleat/Tiling-Assistant/issues/375