chore: add jsconfig and types
This will allow for a more pleasant dev experience by adding better
auto-complete and typing. Especially, for custom properties and methods
that are injected into GNOME's objects.

This is probably also the extent of 'typing' I plan to add. I don't
intend to enable `checkJs` nor do I plan to transition to Typescript.